Corrected Files
These files should be the correct most recent files for both the android and the web applications. The files here should replace the files in the development branch and should be run through all tests to ensure everything works as it should. 

Kyle, I have made it so that the html and js files for the android app are the same as for the web.